+++ reprocess.py
@@ -1,3 +1,9 @@
+"""
+This script allows you to manually trigger the processing
+of submissions that have already been decrypted.
+Requires a file path to the decrypted json as an argument.
+"""
+
 import sys
 from json import load
 
@@ -7,6 +13,13 @@ from app.response_processor import ResponseProcessor
 
 
 def reprocess(path_to_decrypted_json):
+    """Restarts processing after the decryption phase
+
+    Parameters
+    ----------
+    path_to_decrypted_json : str
+        The file location of the decrypted survey submission as json
+    """
 
     with open(path_to_decrypted_json) as json_file:
         decrypted_json = load(json_file)
@@ -21,5 +34,6 @@ def reprocess(path_to_decrypted_json):
 
 
 if __name__ == "__main__":
+    # requires a filepath as the one and only argument
     path_to_json = sys.argv[1]
     reprocess(path_to_json)